crawler/DrissionPage/dp项目(hw)/music163_crawler-master/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cssselect==1.2.0
DataRecorder==3.6.2
DownloadKit==2.0.7
DrissionPage==4.1.0.17
et_xmlfile==2.0.0
filelock==3.17.0
frozenlist==1.5.0
idna==3.10
lxml==5.3.1
multidict==6.1.0
numpy==2.2.3
openpyxl==3.1.5
pandas==2.2.3
propcache==0.3.0
psutil==7.0.0
PyExecJS==1.5.1
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
requests-file==2.1.0
six==1.17.0
tldextract==5.1.3
tzdata==2025.1
urllib3==2.3.0
websocket-client==1.8.0
yarl==1.18.3
